Fixes not-above-threshold slipping through on merits of detections with lower thresholds
- For detection rows where some high-threshold classes are not passing,
they end up being accepted by NMS later because of their relatively high
confidence - instead of lower-thresholded classes that meritted them
passing the initial filter
- This change truncates all non-passing class confidences to 0 to avoid
these to slip through.
- E.g. a bollard of 40% might get through because there is also a
passing confidence for temporary sign at say 30%. While this is a high
confidence for the temporary sign, its very low on the bollard - and
likely not to be a bollard but a temporary sign.